Add support for Rails 6.0 and 6.1
This change adds support for Rails 6.0 and 6.1 by increasing the upper bound of the `activesupport` dependency to `< 6.2`.

To ensure tests pass using 6.0 and 6.1, the Travis CI matrix has been updated with `rails6.0.gemfile` and `rails6.1.gemfile`.

Note: Rails 6.x requires `Ruby 2.5.0` or newer, so exclusions to the matrix have been specified for  `rails6.0.gemfile` and `rails6.1.gemfile`.